refactored musicplayer: extracted infoToggles and currentSongInfo
