+++ client/src/components/MusicPlayer.vue
@@ -12,50 +12,7 @@
     >
       Your browser does not support the audio element.
     </audio>
-    <div class="currentSongInfo">
-      <img :src="coverUrl" class="cover" id="image" />
-      <div class="songInfoText">
-        <h3
-          ref="currentSongName"
-          class="currentSongName"
-          @click="findNodesInGraph('song', [currentSong.id])"
-        >
-          <div
-            :class="{
-              innerText: overflows.currentSongName,
-              innerTextShort: !overflows.currentSongName,
-            }"
-          >
-            {{ currentSong.name
-            }}<span id="textExtension" v-if="overflows.currentSongName">{{
-              currentSong.name
-            }}</span>
-          </div>
-        </h3>
-        <div
-          class="currentArtistName"
-          ref="currentArtistName"
-          @click="
-            findNodesInGraph(
-              'artist',
-              currentSong.artists.map((artist) => artist.id),
-            )
-          "
-        >
-          <div
-            :class="{
-              innerText: overflows.currentArtistName,
-              innerTextShort: !overflows.currentArtistName,
-            }"
-          >
-            {{ getArtists()
-            }}<span id="textExtension" v-if="overflows.currentArtistName">{{
-              getArtists()
-            }}</span>
-          </div>
-        </div>
-      </div>
-    </div>
+    <CurrentSongInfo />
     <div class="player">
       <div class="controls">
         <button
@@ -86,26 +43,6 @@
         ></slider>
       </div>
     </div>
-    <div id="queue">
-      <div id="buttons">
-        <button
-          id="queueButton"
-          small
-          :class="{ activated: queueDisplay }"
-          @click="setQueueVisibility(!queueDisplay)"
-        >
-          <v-icon icon="mdi-playlist-music" />
-        </button>
-        <button
-          id="infoButton"
-          small
-          :class="{ activated: nodeInfoDisplay }"
-          @click="setNodeInfoVisibility(!nodeInfoDisplay)"
-        >
-          <v-icon icon="mdi-information" />
-        </button>
-      </div>
-    </div>
 
     <div class="volume">
       <v-icon>mdi-volume-minus</v-icon>
@@ -123,9 +60,8 @@
 
 <script>
 import { mapState, mapActions } from "vuex";
-import logo from "@/assets/logo.png";
-import { searchGraph } from "@/assets/js/graphHelper.js";
 import Slider from "./Slider.vue";
+import CurrentSongInfo from "./CurrentSongInfo.vue";
 
 export default {
   data: () => ({
@@ -135,40 +71,18 @@ export default {
     start: 0,
     end: 2,
     isSeeking: false,
-    overflows: {
-      currentSongName: false,
-      currentArtistName: false,
-    },
   }),
   components: {
     Slider,
+    CurrentSongInfo,
   },
   computed: {
     ...mapState({
       songUrl: (state) => state.music_player.currentSong.preview_url,
-      queueDisplay: (state) => state.visibleItems.queueDisplay,
-      nodeInfoDisplay: (state) => state.visibleItems.nodeInfo,
     }),
-    currentSong: {
-      get() {
-        return this.$store.state.music_player.currentSong;
-      },
-      set(value) {},
-    },
-    coverUrl() {
-      if (this.currentSong.images.length > 0)
-        return this.currentSong.images[0].url;
-      else return logo;
-    },
   },
   methods: {
-    ...mapActions([
-      "playNextInQueue",
-      "playPreviousInQueue",
-      "setQueueVisibility",
-      "setNodeInfoVisibility",
-      "fitGraphToNodes",
-    ]),
+    ...mapActions(["playNextInQueue", "playPreviousInQueue"]),
     StartStop: function () {
       if (this.isPlaying) {
         this.$refs.player.pause();
@@ -199,51 +113,8 @@ export default {
     PrevSong: function () {
       this.playPreviousInQueue();
     },
-    getArtists: function () {
-      return this.currentSong.artists
-        ? this.currentSong.artists.map((artist) => artist.name).join(", ")
-        : "";
-    },
-    checkOverflow: function (reference) {
-      return (
-        (reference ? reference.scrollWidth : null) >
-        (reference ? reference.offsetWidth : null)
-      );
-    },
-    findNodesInGraph: function (type, sids) {
-      const rootState = this.$store.state;
-      const result = sids.flatMap((sid) => {
-        return searchGraph(
-          {
-            nodeType: type,
-            valid: true,
-            errors: [],
-            attributes: [
-              { attributeData: sid, attributeSearch: "sid", operator: "=" },
-            ],
-          },
-          rootState,
-        );
-      });
-      this.fitGraphToNodes(result);
-    },
   },
   mounted: function () {},
-  watch: {
-    currentSong: function () {
-      this.overflows.currentSongName = false;
-      this.overflows.currentArtistName = false;
-
-      this.$nextTick(function () {
-        this.overflows.currentSongName = this.checkOverflow(
-          this.$refs.currentSongName,
-        );
-        this.overflows.currentArtistName = this.checkOverflow(
-          this.$refs.currentArtistName,
-        );
-      });
-    },
-  },
 };
 </script>
 
@@ -255,74 +126,32 @@ export default {
   border-radius: 5px;
   padding-left: 0.5rem;
   padding-right: 0.5rem;
-
-  margin-left: 25%;
-  margin-bottom: 2%;
-  width: 50%;
-
   display: flex;
   gap: 1rem;
   align-items: center;
   justify-content: center;
 }
-.cover {
-  height: 3rem;
-  width: 3rem;
-  object-fit: contain;
-}
-.currentSongInfo {
-  display: flex;
-  cursor: pointer;
-  gap: 0.5rem;
-  align-items: center;
-}
 
-.currentSongInfo .songInfoText {
-  text-overflow: ellipsis;
-  overflow: hidden;
-  white-space: nowrap;
-}
-
-#queueButton {
-  background-color: #da6a1dff;
-}
-#queueButton:hover {
-  background-color: #1dcdda66;
-}
-#infoButton {
-  background-color: #da6a1dff;
-}
-#infoButton:hover {
-  background-color: #1dcdda66;
-}
-#infoButton .activated {
-  background-color: #1dcddaff;
-}
-#queueButton .activated {
-  background-color: #1dcddaff;
-}
-.currentSongName {
-  overflow: hidden;
-}
-.currentArtistName {
-  overflow: hidden;
-}
-.currentSongName:hover .innerText {
-  animation: marquee 12s linear infinite;
-}
-.currentArtistName:hover .innerText {
-  animation: marquee 12s linear infinite;
-}
-input {
-  color: white;
+@media screen and (max-width: 500px) {
+  .card {
+    flex-direction: column;
+    flex-shrink: 2;
+  }
+  .player {
+    width: 100%;
+  }
+  .currentSongInfo {
+    width: 100%;
+  }
 }
 
 .player {
-  width: 50%;
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 0.3rem;
+  min-width: 75px;
+  flex-grow: 1;
 }
 .sliderBox {
   width: 100%;
